PID DOCUMENTATION
Here is the rough draft of the PID documentation. Please correct any spelling and punctuation, and have a further look at what other team members have contributed towards the document. If any changes need to be made please notify the rest of the team using the commit changes section. Thanks.